Move screensavermanager to scripts
