+++ src/scripts/screensavermanager.js
@@ -1,8 +1,8 @@
 import { Events } from 'jellyfin-apiclient';
 import { playbackManager } from '../components/playback/playbackmanager';
 import { pluginManager } from '../components/pluginManager';
-import inputManager from '../scripts/inputManager';
-import * as userSettings from '../scripts/settings/userSettings';
+import inputManager from './inputManager';
+import * as userSettings from './settings/userSettings';
 import ServerConnections from '../components/ServerConnections';
 import './screensavermanager.scss';
